requirements.txt
streamlit
python-dotenv
langchain
langchain-upstage
transformers
torch